requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
alembic==1.13.2
    # via triangler-fastapi (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   starlette
    #   watchfiles
cffi==1.16.0
    # via cryptography
click==8.1.7
    # via uvicorn
cryptography==43.0.0
    # via python-jose
ecdsa==0.19.0
    # via python-jose
fastapi==0.112.0
    # via triangler-fastapi (pyproject.toml)
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.7
    # via anyio
jinja2==3.1.4
    # via starlette-admin
loguru==0.7.2
    # via triangler-fastapi (pyproject.toml)
mako==1.3.5
    # via alembic
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
numpy==2.0.1
    # via scipy
pillow==10.4.0
    # via
    #   triangler-fastapi (pyproject.toml)
    #   qrcode
pyasn1==0.6.0
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via
    #   triangler-fastapi (pyproject.toml)
    #   fastapi
pydantic-core==2.20.1
    # via pydantic
pypng==0.20220715.0
    # via qrcode
python-dotenv==1.0.1
    # via uvicorn
python-jose==3.3.0
    # via triangler-fastapi (pyproject.toml)
python-multipart==0.0.9
    # via starlette-admin
pyyaml==6.0.1
    # via uvicorn
qrcode==7.4.2
    # via triangler-fastapi (pyproject.toml)
rsa==4.9
    # via python-jose
scipy==1.14.0
    # via triangler-fastapi (pyproject.toml)
six==1.16.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.31
    # via
    #   triangler-fastapi (pyproject.toml)
    #   alembic
starlette==0.37.2
    # via
    #   fastapi
    #   starlette-admin
starlette-admin==0.14.1
    # via triangler-fastapi (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   qrcode
    #   sqlalchemy
uvicorn==0.30.5
    # via triangler-fastapi (pyproject.toml)
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
wheel==0.43.0
    # via triangler-fastapi (pyproject.toml)
